// src/generated/econt/types.gen.ts
export type Address = {
    id?: number;
    city?: City;
    fullAddress?: string;
    fullAddressEn?: string;
    quarter?: string;
    street?: string;
    num?: string;
    other?: string;
    location?: GeoLocation;
    zip?: string;
    hezid?: string;
};

/**
 * City served by Econt Express. Required fields for valid city - ID or name + post code (if the City is outside Bulgaria, country is required)
 *
 */
export type City = {
    id?: number;
    country?: Country;
    postCode?: string;
    name?: string;
    nameEn?: string;
    regionName?: string;
    regionNameEn?: string;
    phoneCode?: string;
    location?: GeoLocation;
    expressCityDeliveries?: boolean;
    monday?: boolean;
    tuesday?: boolean;
    wednesday?: boolean;
    thursday?: boolean;
    friday?: boolean;
    saturday?: boolean;
    sunday?: boolean;
    serviceDays?: number;
    zoneId?: number;
    zoneName?: string;
    zoneNameEn?: string;
    servingOffices?: Array<ServingOfficeElement>;
};

/**
 * Country. Required fields for valid country - ID or code (code2 or code3)
 *
 */
export type Country = {
    id?: number;
    code2?: string;
    code3?: string;
    name?: string;
    nameEn?: string;
    isEU?: boolean;
};

/**
 * A geographic location
 *
 */
export type GeoLocation = {
    latitude?: number;
    longitude?: number;
    confidence?: number;
};
// ...
export type ShipmentType = 'document' | 'pack' | 'pallet' | 'cargo' | 'documentpallet' | 'big_letter' | 'small_letter' | 'money_transfer' | 'pp';
// ...
export type Office = {
    id?: number;
    code?: string;
    isMPS?: boolean;
    isAPS?: boolean;
    name?: string;
    nameEn?: string;
    phones?: Array<string>;
    emails?: Array<string>;
    address?: Address;
    info?: string;
    currency?: string;
    language?: string;
    normalBusinessHoursFrom?: string;
    normalBusinessHoursTo?: string;
    halfDayBusinessHoursFrom?: string;
    halfDayBusinessHoursTo?: string;
    shipmentTypes?: Array<ShipmentType>;
    partnerCode?: string;
    hubCode?: string;
    hubName?: string;
    hubNameEn?: string;
    isDrive?: boolean;
};